third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/at-container-serialization-expected.txt

This is a testharness.js-based test.
[FAIL] @container conditionText serialization: (width),(height) ,--foo ,--bar
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
